Move all sponge to new module.
