+++ src/lib.rs
@@ -55,8 +55,6 @@ mod round_numbers;
 
 /// Sponge
 pub mod sponge;
-pub mod sponge_api;
-pub mod sponge_circuit;
 
 /// Hash types and domain separation tags.
 pub mod hash_type;
